+++ ios/Classes/Model/Domain/BlikPayment.swift
@@ -6,10 +6,10 @@ struct BlikPayment: Transaction {
 
     let amount: Double
     let description: String
+    let hiddenDescription: String?
     let payerContext: PayerContext?
     let token: String?
     let alias: String?
     let paymentChannel: Headless.Models.PaymentChannel
     let callbacks: CallbacksConfiguration
-    let hiddenDescription: String?
 }