+++ ios/Classes/Model/Domain/CardPayment.swift
@@ -6,10 +6,10 @@ struct CardPayment: Transaction {
 
     let amount: Double
     let description: String
+    let hiddenDescription: String?
     let payerContext: PayerContext?
     let paymentChannel: Headless.Models.PaymentChannel
     let card: Headless.Models.Card?
     let cardToken: Headless.Models.CardToken?
     let callbacks: CallbacksConfiguration
-    let hiddenDescription: String?
 }