+++ ios/Classes/Model/Domain/DigitalWalletPayment.swift
@@ -6,9 +6,9 @@ struct DigitalWalletPayment: Transaction {
 
     let amount: Double
     let description: String
+    let hiddenDescription: String?
     let payerContext: PayerContext?
     let paymentChannel: Headless.Models.PaymentChannel
     let token: String
     let callbacks: CallbacksConfiguration
-    let hiddenDescription: String?
 }