+++ ios/Classes/Model/Transportation/Transportation+PaymentDetails.swift
@@ -6,5 +6,6 @@ extension Transportation {
 
         let amount: Double
         let description: String
+        let hiddenDescription: String?
     }
 }