+++ ios/Classes/Model/Transportation/Transportation+SingleTransaction.swift
@@ -6,6 +6,7 @@ extension Transportation {
 
         let amount: Double
         let description: String
+        let hiddenDescription: String?
         let payerContext: PayerContext?
     }
 }