+++ ios/Classes/Model/Transportation/Transportation+TokenPayment.swift
@@ -7,6 +7,7 @@ extension Transportation {
         let cardToken: String
         let amount: Double
         let description: String
+        let hiddenDescription: String?
         let payer: Payer?
     }
 }